camera collisions & buildings
-stopped camera under floor
-stopped camera under mtn
-stopped camera under crater
-stopped camera behind limits
-added building rendering
-added buildingColour class
-added bench
-added blueprint

+++ startScreen.cpp
@@ -323,7 +323,7 @@ void startScreen::begin(){
 	// loading text
 	loadingText = createText();
 	allTexts[loadingText].fontPath = "assets/fonts/zekton.ttf";
-	allTexts[loadingText].fontSize = (float) display_x / 30.0f;
+	allTexts[loadingText].fontSize = (int)((float)display_x / 30.0f);
 	allTexts[loadingText].displayedText = "Loading...";
 	allTexts[loadingText].position = vec2(display_x * 0.425f, display_y * 0.45f);
 	allTexts[loadingText].active = false;
